Fix many bugs, dockerize-apps, delete liquibase
