+++ authentication/src/main/java/com/moviescloud/authentication/AuthenticationApplication.java
@@ -2,10 +2,12 @@ package com.moviescloud.authentication;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
 @ComponentScan(basePackages = {"com.moviescloud.common", "com.moviescloud.authentication"})
+@EntityScan({"com.moviescloud.movies.entities", "com.moviescloud.common.entities"})
 public class AuthenticationApplication {
     public static void main(String[] args) {
         SpringApplication.run(AuthenticationApplication.class, args);
