+++ common/pom.xml
@@ -7,9 +7,26 @@
     <parent>
         <groupId>com.movies-cloud</groupId>
         <artifactId>movies</artifactId>
+        <relativePath>../pom.xml</relativePath>
         <version>1.0</version>
     </parent>
+
     <artifactId>common</artifactId>
     <name>common</name>
+    <packaging>jar</packaging>
+
+    <build>
+        <plugins>
+            <!-- Отключение spring-boot-maven-plugin для этого модуля -->
+            <plugin>
+                <groupId>org.springframework.boot</groupId>
+                <artifactId>spring-boot-maven-plugin</artifactId>
+                <configuration>
+                    <skip>true</skip>
+                </configuration>
+            </plugin>
+        </plugins>
+    </build>
+
 
 </project>
