+++ common/src/main/java/com/moviescloud/common/entities/Country.java
@@ -1,4 +1,4 @@
-package com.moviescloud.movies.entities;
+package com.moviescloud.common.entities;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -15,12 +15,11 @@ import jakarta.validation.constraints.Size;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class Profession {
+public class Country {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-
     @NotBlank
-    @Size(max = 256)
-    private String professionName;
+    @Size(min = 3, max = 300)
+    private String name;
 }
