+++ common/src/main/java/com/moviescloud/common/entities/Review.java
@@ -1,4 +1,4 @@
-package com.moviescloud.movies.entities;
+package com.moviescloud.common.entities;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.AllArgsConstructor;
@@ -7,44 +7,34 @@ import lombok.NoArgsConstructor;
 
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.PastOrPresent;
 import jakarta.validation.constraints.Size;
 import java.util.Date;
-import java.util.List;
 
 @Entity
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class Person {
-
+public class Review {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private long id;
 
     @NotBlank
-    @Size(min = 5, max = 256)
-    private String nameRu;
+    @Size(max = 2500)
+    private String title;
 
     @NotBlank
-    @Size(min = 5, max = 256)
-    private String nameEn;
-
-    @NotBlank
-    @Size(max = 256)
-    private String posterUrl;
+    @Size(max = 640000)
+    @Column(columnDefinition="TEXT")
+    private String description;
 
     @PastOrPresent
-    private Date birthday;
+    private Date createdAt;
 
-    @NotBlank
-    @OneToOne
-    private Country birthPlace;
-
-    @ManyToMany
     @JsonIgnore
-    private List<Movie> movies;
-
-    @OneToMany
-    private List<Profession> professions;
+    @ManyToOne
+    @NotNull
+    private User author;
 }
