+++ common/src/main/resources/application.properties
@@ -4,5 +4,5 @@ spring.datasource.password=password
 spring.datasource.driver-class-name=org.postgresql.Driver
 user.service.url=http://user-service/api/v1/users
 privilege.service.url=http://user-service/api/v1/privileges
-jwt.secret=very_long_text
-jwt.validity=43200000
+#user.service.url=http://localhost:8081/api/v1/users
+#privilege.service.url=http://localhost:8081/api/v1/privileges
