+++ movies-main/pom.xml
@@ -5,10 +5,12 @@
     <parent>
         <groupId>com.movies-cloud</groupId>
         <artifactId>movies</artifactId>
+        <relativePath>../pom.xml</relativePath>
         <version>1.0</version>
     </parent>
     <artifactId>movies-main</artifactId>
     <name>movies-main</name>
+
     <dependencies>
         <dependency>
             <groupId>com.movies-cloud</groupId>
@@ -17,4 +19,5 @@
             <scope>compile</scope>
         </dependency>
     </dependencies>
+
 </project>
