+++ movies-main/src/main/java/com/moviescloud/movies/repositories/CountryRepository.java
@@ -1,6 +1,6 @@
 package com.moviescloud.movies.repositories;
 
-import com.moviescloud.movies.entities.Country;
+import com.moviescloud.common.entities.Country;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
