+++ movies-main/src/main/java/com/moviescloud/movies/repositories/GenreRepository.java
@@ -1,6 +1,6 @@
 package com.moviescloud.movies.repositories;
 
-import com.moviescloud.movies.entities.Genre;
+import com.moviescloud.common.entities.Genre;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
