+++ movies-main/src/main/java/com/moviescloud/movies/repositories/MovieRepository.java
@@ -1,9 +1,9 @@
 package com.moviescloud.movies.repositories;
 
-import com.moviescloud.movies.entities.Country;
-import com.moviescloud.movies.entities.Genre;
-import com.moviescloud.movies.entities.Movie;
-import com.moviescloud.movies.entities.Type;
+import com.moviescloud.common.entities.Country;
+import com.moviescloud.common.entities.Genre;
+import com.moviescloud.common.entities.Movie;
+import com.moviescloud.common.entities.Type;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
