+++ movies-main/src/main/java/com/moviescloud/movies/repositories/PersonRepository.java
@@ -1,6 +1,6 @@
 package com.moviescloud.movies.repositories;
 
-import com.moviescloud.movies.entities.Person;
+import com.moviescloud.common.entities.Person;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
