+++ movies-main/src/main/java/com/moviescloud/movies/repositories/ReviewRepository.java
@@ -1,6 +1,6 @@
 package com.moviescloud.movies.repositories;
 
-import com.moviescloud.movies.entities.Review;
+import com.moviescloud.common.entities.Review;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
