+++ movies-main/src/main/java/com/moviescloud/movies/repositories/TypeRepository.java
@@ -1,6 +1,6 @@
 package com.moviescloud.movies.repositories;
 
-import com.moviescloud.movies.entities.Type;
+import com.moviescloud.common.entities.Type;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TypeRepository extends JpaRepository<Type, Long> {
