+++ movies-main/src/main/java/com/moviescloud/movies/services/impl/MovieServiceImpl.java
@@ -1,9 +1,9 @@
 package com.moviescloud.movies.services.impl;
 
-import com.moviescloud.movies.entities.Country;
-import com.moviescloud.movies.entities.Genre;
-import com.moviescloud.movies.entities.Movie;
-import com.moviescloud.movies.entities.Type;
+import com.moviescloud.common.entities.Country;
+import com.moviescloud.common.entities.Genre;
+import com.moviescloud.common.entities.Movie;
+import com.moviescloud.common.entities.Type;
 import com.moviescloud.common.exceptions.ResourceNotFoundException;
 import com.moviescloud.movies.repositories.MovieRepository;
 import com.moviescloud.movies.services.IMovieService;
