+++ movies-main/src/main/java/com/moviescloud/movies/services/impl/ReviewServiceImpl.java
@@ -1,6 +1,6 @@
 package com.moviescloud.movies.services.impl;
 
-import com.moviescloud.movies.entities.Review;
+import com.moviescloud.common.entities.Review;
 import com.moviescloud.common.exceptions.ResourceNotFoundException;
 import com.moviescloud.movies.repositories.ReviewRepository;
 import com.moviescloud.movies.services.IReviewService;
