--- movies-main/src/main/resources/db/changelog/db.changelog-master.yaml
@@ -1,3 +1,0 @@
-databaseChangeLog:
-  - include:
-      file: db/changelog/changes/01-initial-schema.yaml
