+++ pom.xml
@@ -2,24 +2,28 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
+
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
 		<version>3.4.3</version>
 		<relativePath/>
 	</parent>
+
 	<modules>
 		<module>movies-main</module>
 		<module>authentication</module>
 		<module>user-service</module>
 		<module>common</module>
 	</modules>
+
 	<groupId>com.movies-cloud</groupId>
 	<artifactId>movies</artifactId>
 	<version>1.0</version>
 	<name>movies</name>
 	<packaging>pom</packaging>
 	<description>Spring boot API with information a movies</description>
+
 	<properties>
 		<java.version>21</java.version>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
@@ -27,10 +31,10 @@
 		<postgresql.version>42.7.4</postgresql.version>
 		<openapi.version>1.8.0</openapi.version>
 		<com.auth0.version>4.2.2</com.auth0.version>
-		<liquibase.version>4.31.1</liquibase.version>
 		<jakarta.servlet.api.version>6.1.0</jakarta.servlet.api.version>
 		<lombok.version>1.18.36</lombok.version>
 	</properties>
+
 	<build>
 		<plugins>
 			<plugin>
@@ -52,6 +56,7 @@
 			</plugin>
 		</plugins>
 	</build>
+
 	<dependencies>
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
@@ -75,12 +80,6 @@
 			<artifactId>postgresql</artifactId>
 			<version>${postgresql.version}</version>
 			<scope>runtime</scope>
-		</dependency>
-
-		<dependency>
-			<groupId>org.liquibase</groupId>
-			<artifactId>liquibase-core</artifactId>
-			<version>${liquibase.version}</version>
 		</dependency>
 
 		<dependency>
