+++ user-service/pom.xml
@@ -2,13 +2,17 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
+
     <parent>
         <groupId>com.movies-cloud</groupId>
         <artifactId>movies</artifactId>
+        <relativePath>../pom.xml</relativePath>
         <version>1.0</version>
     </parent>
+
     <artifactId>user-service</artifactId>
     <name>user-service</name>
+
     <dependencies>
         <dependency>
             <groupId>com.movies-cloud</groupId>
@@ -17,4 +21,5 @@
             <scope>compile</scope>
         </dependency>
     </dependencies>
+
 </project>
