+++ user-service/src/main/java/com/moviescloud/userservice/UserServiceApplication.java
@@ -2,10 +2,12 @@ package com.moviescloud.userservice;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
 @ComponentScan(basePackages = {"com.moviescloud.common", "com.moviescloud.userservice"})
+@EntityScan({"com.moviescloud.movies.entities", "com.moviescloud.common.entities"})
 public class UserServiceApplication {
     public static void main(String[] args) {
         SpringApplication.run(UserServiceApplication.class, args);
